requirements.txt
streamlit>=1.32.0
PyPDF2>=3.0.1
spacy>=3.7.2
pdf2image
pytesseract
Pillow
